refactor: move ConfidentialResponsesExportEntry and related functions to studyutils package
