+++ services/management-api/apihandlers/study-management.go
@@ -2830,48 +2830,6 @@ type ConfidentialResponsesExportQuery struct {
 	KeyFilter      string   `json:"keyFilter"`
 }
 
-func parseSlots(respItem *studyTypes.ResponseItem, slotKey string) map[string]string {
-	parsedResp := map[string]string{}
-	if respItem == nil {
-		return parsedResp
-	}
-
-	currentSlotKey := slotKey + "." + respItem.Key
-	if strings.HasSuffix(slotKey, "-") {
-		currentSlotKey = slotKey + respItem.Key
-	}
-
-	if len(respItem.Items) == 0 {
-		parsedResp[currentSlotKey] = respItem.Value
-		return parsedResp
-	}
-
-	for _, subItem := range respItem.Items {
-		r := parseSlots(subItem, currentSlotKey)
-		for k, v := range r {
-			parsedResp[k] = v
-		}
-	}
-	return parsedResp
-}
-
-func confidentialResponseExport(resp studyTypes.SurveyResponse, realPID string) map[string]string {
-	parsedResp := map[string]string{
-		"participantID": realPID,
-	}
-
-	for _, r := range resp.Responses {
-		slotKey := r.Key + "-"
-
-		slots := parseSlots(r.Response, slotKey)
-		for k, v := range slots {
-			parsedResp[k] = v
-		}
-	}
-
-	return parsedResp
-}
-
 func (h *HttpEndpoints) getConfidentialResponses(c *gin.Context) {
 	token := c.MustGet("validatedToken").(*jwthandling.ManagementUserClaims)
 
@@ -2903,7 +2861,7 @@ func (h *HttpEndpoints) getConfidentialResponses(c *gin.Context) {
 	idMappingMethod := study.Configs.IdMappingMethod
 	globalSecret := h.globalStudySecret
 
-	results := []map[string]string{}
+	results := []studyutils.ConfidentialResponsesExportEntry{}
 
 	for _, pID := range query.ParticipantIDs {
 		// confidentialID := studyTypes.GetConfidentialParticipantID(token.InstanceID, studyKey, pID)
@@ -2921,7 +2879,7 @@ func (h *HttpEndpoints) getConfidentialResponses(c *gin.Context) {
 		}
 
 		for _, r := range responses {
-			results = append(results, confidentialResponseExport(r, pID))
+			results = append(results, studyutils.PrepConfidentialResponseExport(r, pID, nil)...)
 		}
 	}
